# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "基于 Gin 和 GORM 构建的高性能社交网络平台接口文档",
        "title": "社交网络平台API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API 支持团队",
            "url": "http://www.swagger.io/support",
            "email": "caribdis@163.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/messages": {
            "get": {
                "description": "获取与指定好友的聊天记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "获取聊天记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "好友ID",
                        "name": "friend_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "开始时间戳(默认7天前)",
                        "name": "start_time",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "结束时间戳(默认当前时间)",
                        "name": "end_time",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否标记为已读",
                        "name": "mark_read",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "是否获取全部历史消息",
                        "name": "history",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"messages\":[{\"from\":发送者ID,\"to\":接收者ID,\"direct\":消息发送方向 ,\"created_at\":时间,\"content\":\"内容\",\"avatar_url\":\"头像URL\",\"hide_time\":是否隐藏时间}]}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "错误信息",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "向指定好友发送文本消息(图片和文件消息请使用/messages/image和/messages/file路由)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "发送文本消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文本消息参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamTextReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"timestamp\":时间戳,\"receiver_id\":接收者ID,\"status\":\"sent\"}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "错误信息",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/messages/file": {
            "post": {
                "description": "向指定好友发送文件消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "发送文件消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "文件消息参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamFileReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"timestamp\":时间戳,\"receiver_id\":接收者ID,\"status\":\"sent\"}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "错误信息",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/messages/image": {
            "post": {
                "description": "向指定好友发送图片消息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "发送图片消息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "图片消息参数",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.ParamImageReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"timestamp\":时间戳,\"receiver_id\":接收者ID,\"status\":\"sent\"}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "错误信息",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/messages/unread": {
            "get": {
                "description": "获取所有好友的未读消息数",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "消息"
                ],
                "summary": "获取未读消息数",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"counts\":{\"好友ID\":未读数量}}",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    },
                    "400": {
                        "description": "错误信息",
                        "schema": {
                            "$ref": "#/definitions/models.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "models.ParamFileReq": {
            "type": "object",
            "required": [
                "content",
                "to"
            ],
            "properties": {
                "content": {
                    "description": "文件URL",
                    "type": "string"
                },
                "name": {
                    "description": "文件名",
                    "type": "string"
                },
                "size": {
                    "description": "文件大小(字节)",
                    "type": "integer"
                },
                "to": {
                    "type": "string",
                    "example": "0"
                },
                "type": {
                    "description": "文件类型",
                    "type": "string"
                }
            }
        },
...
        "models.ParamImageReq": {
            "type": "object",
            "required": [
                "content",
                "to"
            ],
            "properties": {
                "content": {
                    "description": "图片URL",
                    "type": "string"
                },
                "height": {
                    "description": "图片高度(像素)",
                    "type": "integer"
                },
                "to": {
                    "type": "string",
                    "example": "0"
                },
                "width": {
                    "description": "图片宽度(像素)",
                    "type": "integer"
                }
            }
        },
...
        "models.ParamTextReq": {
            "type": "object",
            "required": [
                "content",
                "to"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "to": {
                    "type": "string",
                    "example": "0"
                }
            }
        },
...
        "models.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "状态码",
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "description": "业务数据（如UID字符串）"
                },
                "msg": {
                    "description": "消息描述",
                    "type": "string",
                    "example": "success"
                }
            }
        }
    }
}